ml-service/samples/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
mysql-connector-python>=8.3.0
Pillow>=10.0.0
